requirements.txt
coverage==5.1
Django==3.0.4
djangorestframework==3.11.0